+++ src/ClipboardMenu.hpp
@@ -24,6 +24,4 @@ public:
 
     float getButtonScale() const;
     int getButtonOpacity() const;
-
-    bool isRTL() const;
 };