+++ changelog.md
@@ -1,3 +1,6 @@
+# v1.0.2
+- Removed clipboard UI auto-align on focus
+
 # v1.0.1
 - Crash fixes
 - Removed toggle setting `Right to Left`